linux-mainline: Revert "random: fix crng_ready() test" patch
This reverts commit 43838a23a05fbd13e47d750d3dfd77001536dd33.

The reason for reverting this commit is that entropy gets acquired very
hard on Orange Pi Plus2.

The hardware crypto module is not supported yet for H3 SoC in this kernel version.
It is expected to be included in 4.18 kernel.

As NetworkManager is depending on randomness to succesfully set its connections, if we don't revert
this commit, the board can get blocked for minutes until it succesfully connects to a network.

This patch should be reverted immediately when upgrading to a kernel where hw crypto is supported.

Signed-off-by: Sebastian Panceac <sebastian@resin.io>