u-boot_2018.09.bbapp:Add resin-specific-env-integration-kconfig.patch
This patch is a rework of the meta-resin patch with the same
function because the meta-resin patch fails to apply due to
different source context in this u-boot version.

The 0001-nanopi_neo_air_defconfig-Enable-eMMC-support.patch
from meta-sunxi was failing to apply and it needed to be
removed and replaced with the one in this
layer (warrior update).

Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>