u-boot_2018.09.bbappend: Rename SRC_URI_remove patch
A patch removed with SRC_URI_remove and re-worked was not
applying (no output error, the behavior is as it doesn't
exist), because the re-worked patch had the same name
as the removed one.

Remove 0001-nanopi_neo_air_defconfig-Enable-eMMC-support.patch.
This patch exists in meta-sunxi and it applies over
versions of u-boot.2018.09

Changelog-entry: Rename re-worked patch to have it applied
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>