linux-mainline_%.bbappend: Add kernel config for configfs
Customer requested i2c and UART enabled on the orange pi
zero. The directory /sys/kernel/config/device-tree
which is used to add overlays at runtime is not present
in the file system.
This patch enables configfs and generation of the needed
directory.

Changelog-entry: Add configs in the kernel to enable configfs
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>