linux-mainline: Remove the heap of patches from armbian
These patches were too many to actually be easy to follow when we
needed to update the kernel. So let's remove them altogether and just
rely on the kernel recipe from the BSP. In case we'll need patches from
armbian in the future we will just port here the ones we strictly need.

Signed-off-by: Florin Sarbu <florin@balena.io>