linux-firmware: Fix wifi for BananaPi M1+
Wifi did not work on the BananaPi M1+ before this
commit, so we have to install the specific bcm43362-sdio file
for this device type.

This firmware configuration file is specific to the M1+ and comes from
https://github.com/armbian/firmware/blob/master/ap6210/nvram_ap6210.txt
and has the checksum fb1ab0d5148e3f58ce94598665828713.

Changelog-entry: linux-firmware: Fix wifi for BananaPi M1+
Signed-off-by: Alexandru Costache <alexandru@balena.io>